+++ src/main/java/com/github/swrirobotics/bags/BagController.java
@@ -33,10 +33,7 @@ package com.github.swrirobotics.bags;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.swrirobotics.bags.persistence.Bag;
 import com.github.swrirobotics.bags.reader.exceptions.BagReaderException;
-import com.github.swrirobotics.support.web.BagList;
-import com.github.swrirobotics.support.web.BagUpdateStatus;
-import com.github.swrirobotics.support.web.ExtJsFilter;
-import com.github.swrirobotics.support.web.ExtJsFilterFormatter;
+import com.github.swrirobotics.support.web.*;
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
 import org.slf4j.Logger;
@@ -131,6 +128,15 @@ public class BagController {
         }
 
         return status;
+    }
+
+    @RequestMapping("/treenode")
+    public BagTreeNode[] getTreeNode(@RequestParam String node) throws IOException {
+        myLogger.info("getTreeNode: " + node);
+
+        List<BagTreeNode> nodes = myBagService.getTreePath(node);
+
+        return myBagService.getTreePath(node).toArray(new BagTreeNode[nodes.size()]);
     }
 
     @InitBinder